Added prototype scripts for dataset creation
These are 3 prototype scripts that are used in the dataset creation for our Verilog translation LLM. They work as follows: 
1. dataset_parse uses the verilog_autocompletion's csv file of github repositories and clones them, then processes the data by removing every non-verilog file
1.a. Ideally this would also delete the cloned repositories, but I have issues doing this on my Windows machine
2. copy_dir_structure copies the file structure from one directory to another. This is a mandatory prerequisite for the third script
3. auth_synth (which is run INSIDE the OSS-CAD-Suite) reads in a verilog file and outputs a synthesized (i.e. elaborated) design into the specified output file
With these 3 scripts, we can gather the raw data needed for an initial dataset. 
However, the data needs to be processed, and perhaps the scripts made more elaborate as to handle error since there isn't really any error-handling in them.